pyproject.toml
[tool.poetry]
name = "hexlet-code"

version = "0.1.0"
description = "gendiff"
authors = ["Evgeny <info@tinamagic.lv>"]
readme = "README.md"

classifiers = ["Topic :: Software Development :: Gendiff"]
repository = "https://github.com/EvgenyAleksov/python-project-50"

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
prompt = "^0.4.1"

pyyaml = "^6.0.1"
[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
